% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{class}
\name{CameraParam-class}
\alias{CameraParam-class}
\title{A class containing the filter parameters}
\description{
A class containing the filter parameters
}
\section{Slots}{

\describe{
\item{\code{cores}}{`numeric(1)` number of cores for parallelization}

\item{\code{polarity}}{`character(1)` "positive" or "negative"}

\item{\code{sigma}}{`numeric(1)` multiplier of the standard deviation}

\item{\code{perfwhm}}{`numeric(1)` percentage of the FWHM}

\item{\code{intval}}{`character(1)` "into", "maxo" or "intb"}

\item{\code{cor_eic_th}}{`numeric(1)` correlation threshold}

\item{\code{pval}}{`numeric(1)` significant correlation threshold}

\item{\code{graph_method}}{`character(1)` method selection for grouping peaks after
correlation analysis into pseudospectra, could be "hcs" or "lpc"}

\item{\code{calcIso}}{`logical(1)` use isotopic relationship for peak grouping}

\item{\code{calcCiS}}{`logical(1)` use correlation inside samples for peak grouping}

\item{\code{calcCaS}}{`logical(1)` use correlation across samples for peak grouping}

\item{\code{maxcharge}}{`numeric(1)` max ion charge}

\item{\code{maxiso}}{`numeric(1)` max isotopologues}

\item{\code{ppm}}{`numeric(1)` ppm tolerance}

\item{\code{mzabs}}{`numeric(1)` mDa tolerance}

\item{\code{minfrac}}{`numeric(1)` percentage number of samples which must satisfy
12C/13C rule}

\item{\code{multiplier}}{`numeric(1)` max number n in [nM+H]+}

\item{\code{max_peaks}}{`numeric(1)` max how much peaks per thread}

\item{\code{rules}}{`dataframe` with columns:
\itemize{
    \item name `character` name of the adduct
    \item nmol `numeric` number n in [nM+H]+
    \item charge `numeric` charge of the adduct
    \item massdiff `numeric` mass difference with the adduct
    \item oidscore `numeric` adduct with same kation have same score
    (ex: [M+H]+ and [2M+H]+)
    \item quasi `numeric` valid a pc group if one flagged with at least one
    adduct at 1
    \item ips `numeric` rule score, if one peak can be explained,
    only its pc group can be kept
}}
}}

