% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{CameraParam}
\alias{CameraParam}
\title{Create CameraParam object}
\usage{
CameraParam(
  ann_params,
  cores = 1,
  sigma = 6,
  perfwhm = 0.6,
  cor_eic_th = 0.75,
  pval = 0.05,
  graph_method = "hcs"
)
}
\arguments{
\item{ann_params}{`AnnotationParam` object}

\item{cores}{`numeric(1)` number of cores for parallelization}

\item{sigma}{`numeric(1)` multiplier of the standard deviation}

\item{perfwhm}{`numeric(1)` percentage of the FWHM}

\item{cor_eic_th}{`numeric(1)` correlation threshold}

\item{pval}{`numeric(1)` significant correlation threshold}

\item{graph_method}{`character(1)` method selection for grouping peaks after
correlation analysis into pseudospectra, could be "hcs" or "lpc"}
}
\value{
`CameraParam` object
}
\description{
Create CameraParam object, some parameters like `max_iso` will be computed by
 loading the chemical database
This object will store CAMERA parameters
}
\examples{
\dontrun{
ann_params <- AnnotationParam(
   da_tol = 0.015,
   rt_tol = 10,
   abd_tol = 25,
   adduct_names = c(
       "[M+Na]+",
       "[M+NH4]+",
       "[M+H-H2O]+",
       "[M+H]+",
       "[M-H]-"
   ),
   instrument = "QTOF_XevoG2-S_R25000@200",
   database = "test",
   cpd_classes = c("LPC", "Cer", "FA")
)
camera_params <- CameraParam(
   ann_param = ann_param,
   cores = 1,
   sigma = 6,
   perfwhm = .6,
   cor_eic_th = .75,
   pval = .05,
   graph_method = "hcs"
)
}
}
