% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{check_ms_process_args}
\alias{check_ms_process_args}
\title{Check all args of "Process"}
\usage{
check_ms_process_args(mzxml_files, cwt_params, obw_params, cores)
}
\arguments{
\item{mzxml_files}{vector of filepaths to the files}

\item{cwt_params}{CentWaveParam object created by xcms, 
if not given it will launch the parameters by default}

\item{obw_params}{ObiwarpParam object created by xcms, 
if not given it will launch the parameters by default}

\item{cores}{numeric number of cores to use for the peakpicking 
(one core = one file peakpicked)}
}
\description{
Check all arguments of the "Process" step. 
     Will raise an error if one argument is incorrect
}
