% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakpicking.R
\name{check_peakpicking_args}
\alias{check_peakpicking_args}
\title{Check peakpicking args}
\usage{
check_peakpicking_args(mzxml_files, cwt_params, cores)
}
\arguments{
\item{mzxml_files}{vector of filepaths to the files}

\item{cwt_params}{CentWaveParam object created by xcms, 
if not given it will launch the parameters by default}

\item{cores}{numeric number of cores to use for the peakpicking 
(one core = one file peakpicked)}
}
\description{
Check peakpicking args. Will raise an error if one argument is incorrect
}
