% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{compare_spectras}
\alias{compare_spectras}
\title{Compare spectras}
\usage{
compare_spectras(
  q_spectra,
  l_spectras,
  da_tol = 0.05,
  abd_tol = 25,
  suffix = "theo"
)
}
\arguments{
\item{q_spectra}{dataframe with columns : 
\itemize{
    \item mz m/z
    \item int or abd for intensity or relative abundance tolerance
}}

\item{l_spectras}{list of dataframe with columns :
\itemize{
    \item mz m/z
    \item int or abd for intensity or relative abundance tolerance
}}

\item{da_tol}{Da tolerance}

\item{abd_tol}{abundance tolerance, at which \% tolerance a peak need to be}

\item{suffix}{which suffix to apply to 
the column names of the library spectras}
}
\value{
list with : 
     \itemize{
         \item score the isotopic score between query & library spectra
         \item deviation_mz the meanned m/z deviation
         \item npeak number of peaks matched in query spectra
         \item spectras list of dataframe consisting of 
             the merge of the query spectra with the corresponding lines 
             of the library spectra
     }
     each item is repeated x times, 
         corresponding to the number of library spectras
}
\description{
Compare a query spectra against library spectras
}
