% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert_file}
\alias{convert_file}
\title{Convert RAW file}
\usage{
convert_file(raw_file, outdir, converter, verbose = TRUE)
}
\arguments{
\item{raw_file}{filepath}

\item{outdir}{dirpath to the directory of the converted file}

\item{converter}{filepath to msconvert executable}

\item{verbose}{boolean disable prints}
}
\value{
filepath to the converted file
}
\description{
Convert one RAW file to a mzML file
Check with MSnbase package if the file can be converted
If the file is a CDF it will copy the file instead
If the conversion failed & the file is a mzML or mzXML 
     it will copy the file instead
}
