% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert_files}
\alias{convert_files}
\title{Convert RAW files}
\usage{
convert_files(raw_files, converter, verbose = TRUE)
}
\arguments{
\item{raw_files}{vector of filepaths}

\item{converter}{filepath to the msonvert executable}

\item{verbose}{boolean disable prints}
}
\value{
vector of filepaths to the mzXML files converted
}
\description{
Convert RAW files with msconvert to mzXML files
}
\examples{
\dontrun{convert_files("testdata/small.RAW", 
     converter = "pwiz/msconvert.exe")}
}
