% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_copy_eic_mzmat}
\alias{db_copy_eic_mzmat}
\title{Duplicate EICs & m/z mat}
\usage{
db_copy_eic_mzmat(db, samples, old_group_ids, new_group_ids)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{samples}{`character` sample names (corresponding column names to copy)}

\item{old_group_ids}{`numeric` the old group IDs of the EICs & m/z mat to
copy}

\item{new_group_ids}{`numeric` the group IDs to assign to the copied EICs &
m/z mat}
}
\description{
Duplicate EICs & m/z matrix & assign them new group IDs
}
