% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_group_id}
\alias{db_get_group_id}
\title{Get Group ID of ion}
\usage{
db_get_group_id(db, name = NULL, row_id = NULL)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{name}{`character(1)` compound name, optional}

\item{row_id}{`numeric(1)` row ID of the ann DataFrame}
}
\value{
`numeric(1)` group ID
}
\description{
Get group ID of an ion. If only the name is given it will return the group ID
 of the referent ion
}
