% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_group_ids}
\alias{db_get_group_ids}
\title{Get all group IDs}
\usage{
db_get_group_ids(db, spectra_ids)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{spectra_ids}{`numeric` spectra IDs}
}
\value{
`numeric` group IDs
}
\description{
Get all group IDs for a batch of spectra IDs
}
