% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_max_ids}
\alias{db_get_max_ids}
\title{Get the highest IDs}
\usage{
db_get_max_ids(db)
}
\arguments{
\item{db}{`SQLiteConnection`}
}
\value{
`numeric(5)` named vector with the highest ID :
\itemize{
   \item feature
   \item group
   \item spectra
   \item pcgroup
   \item cluster
}
}
\description{
Get the highest ID for each table in database
}
