% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_mzmat}
\alias{db_get_mzmat}
\title{Get m/z matrix}
\usage{
db_get_mzmat(db, group_id)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{group_id}{`numeric(1)` group ID}
}
\value{
`DataFrame` with a column rT & multiple columns, each of them
corresponding to a sample with the intensity
}
\description{
Get m/z deviations recorded in the database for peaks with the same group ID
}
