% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_peakgroups}
\alias{db_get_peakgroups}
\title{Get peakgroups}
\usage{
db_get_peakgroups(db, group_ids = NULL)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{group_ids}{`numeric` group IDs}
}
\value{
`DataFrame`, each line correspond to a group annotated by
XCMS and CAMERA
\itemize{
    \item group_id `integer` group ID
    \item pcgroup_id `integer` pcgroup ID
    \item adduct `character` adduct annotation by CAMERA (NULL if absent)
    \item cluster_id `integer` cluster ID
    \item iso `character` could be "M" or "M+*"
    \item mzmed `float` m/z median computed by XCMS
    \item mzmin `float` m/z median born min (not the m/z born min !!)
    \item mzmax `float` m/z median born max (not the m/z born max !!)
    \item rtmed `float` rT median computed by XCMS
    \item rtmin `float` rT median born min (not the rT born min !!)
    \item rtmax `float` rT median born max (not the rT born max !!)
    \item npeaks `integer` number of peaks grouped accross samples
    \item ... `integer` a column for each sample which contain the
    feature ID (row ID from the peaktable)
}
}
\description{
Get peakgroups from database
}
