% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_query}
\alias{db_get_query}
\title{Get result query}
\usage{
db_get_query(db, query, ...)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{query}{`character(1)`}

\item{...}{other parameters passed on the method}
}
\value{
`DataFrame` of the result query
}
\description{
Override the default function `DBI::dbGetQuery` in order to return an empty
dataframe if RSQLite or DBI return the error "no such table"
}
\seealso{
`DBI::dbGetQuery`
}
