% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_row_id}
\alias{db_get_row_id}
\title{Get row ID}
\usage{
db_get_row_id(db, cpd_name)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{cpd_name}{`character(1)` name of the compound}
}
\value{
`numeric(1)` the EIC ID
}
\description{
Get the row ID from the database for a compound according its referent adduct
}
