% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_record_mzdata}
\alias{db_record_mzdata}
\title{Record EICs}
\usage{
db_record_mzdata(db, xset)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{xset}{`xcmsSet` xcms object}
}
\description{
Record EICs and mz for each basepeak and sample
It will be faster than reloading each raw file from database & retrace the
corresponding EIC or the m/z deviations
}
