% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{filter_ms_file}
\alias{filter_ms_file}
\title{Filter MS file}
\usage{
filter_ms_file(ms_file, exp_rt_range, overwrite = TRUE)
}
\arguments{
\item{ms_file}{`OnDiskMSnExp` MSnbase object}

\item{exp_rt_range}{`numeric(2)` expected rT range (in seconds !!!)}

\item{overwrite}{`logical(1)` should overwrite the file if needed ?}
}
\value{
`character(1)` filepath
}
\description{
Filter MS file according a rT range in seconds
don't filter on the m/z dimension (too long & too heavy in flash memory)
}
