% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{get_kendrick_mass}
\alias{get_kendrick_mass}
\title{Compute Kendrick Mass}
\source{
\url{https://doi.org/10.1007/s13361-018-2040-9}
}
\usage{
get_kendrick_mass(mzs)
}
\arguments{
\item{mzs}{`numeric` m/z vector}
}
\value{
`list` with items:
\itemize{
    \item x `numeric` x coordinates (KM)
    \item y `numeric` y coordinates (KMD)
}
}
\description{
Compute Kendrick mass according the base "CH2"
KM : Kendrick Mass
KMD : Kendrick Mass Defect
}
