% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{get_raw_tic}
\alias{get_raw_tic}
\title{Get TIC}
\usage{
get_raw_tic(raw_files, msaccess, polarity = "positive")
}
\arguments{
\item{raw_files}{`character vector` filepaths to the raw files}

\item{msaccess}{`character(1)` filepath to the msaccess executable file}

\item{polarity}{`character(1)` should be "positive" or "negative" only}
}
\value{
`list of DataFrames` each `DataFrame` represent a file and has
 columns :
\itemize{
   \item index
   \item id
   \item event
   \item analyzer
   \item msLevel
   \item rt
   \item int
}
}
\description{
Get TIC from raw files using the msaccess executable file from ProteoWizard
It filters on the polarity of the scans & the ms level (by default 1)
}
