% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Hello world}
\usage{
hello_world()
}
\value{
a list :
		[[1]]
			[1] "foo" "bar"

		[[2]]
			[1] 0 1
}
\description{
Just a test for the package. Will be removed very soon
}
\examples{
hello_world()
}
