% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{load_db_ions}
\alias{load_db_ions}
\title{Get m/z of database}
\usage{
load_db_ions(adduct_names, instrument)
}
\arguments{
\item{adduct_names}{vector of adduct names present in the enviPat list}

\item{instrument}{instrument name for enviPat}
}
\value{
DataFrame with 
\itemize{
     \item id integer ID
     \item formula chemical formula 
     \item name of the compound
     \item rt retention time of the compound
     \item adduct adduct name 
     \item ion_formula ion chemical formula 
     \item charge charge of the ion 
     \item mz m/z 
     \item abd relative abundance 
     \item iso isotopologue annotation 
}
}
\description{
Get all m/z & relative abundance for the internal database
}
