% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\docType{class}
\name{model-class}
\alias{model-class}
\alias{model}
\title{A Reference Class to represent a model in spring}
\description{
Class of object returned by the \code{criteria} or \code{cv.spring} functions.
}

