% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spring-class.R
\docType{methods}
\name{model}
\alias{model}
\alias{model.spring}
\alias{model,spring-method}
\alias{model,list-method}
\title{Picking a model among a spring fit}
\usage{
model(object, lambda1, lambda2=NULL)
}
\arguments{
\item{object}{A list or a single object with class
\code{spring}. Typically, an output of the \code{spring} fitting
function.}

\item{lambda1}{a scalar. The model with the closest \code{lambda1} is picked out from the available fit in \code{object}.}

\item{lambda2}{a scalar. The model with the closest \code{lambda2}
is picked out from the available fit in \code{object}. Useless
(and ignored) if a single \code{spring} object is provided.}
}
\value{
A object with class \code{model}. See the corresponding
documentation and method.
}
\description{
Send a model corresponding to a given couple of \code{lambda1},
\code{lambda2}.  If applied on a spring object, \code{lambda2} is
ignored. if applied on a list of spring object, sleection is made
according both \code{lambda1} and \code{lambda2}.
}
\seealso{
\code{\linkS4class{model}}.
}
