% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ms_process}
\alias{ms_process}
\title{Execute The step "Process"}
\usage{
ms_process(
  mzxml_files,
  cwt_params,
  obw_params = NULL,
  pd_params = NULL,
  ann_params,
  cores = parallel::detectCores(),
  show_pb = TRUE
)
}
\arguments{
\item{mzxml_files}{vector of filepaths to the files}

\item{cwt_params}{CentWaveParam object created by xcms}

\item{obw_params}{ObiwarpParam object created by xcms}

\item{pd_params}{PeakDensityParam object created by xcms}

\item{ann_params}{AnnotationParam object}

\item{cores}{numeric number of cores to use for the peakpicking 
(one core = one file peakpicked)}

\item{show_pb}{boolean print a progress bar or not}
}
\value{
XCMSnExp object
}
\description{
Execute the step "Process" of the workflow.
it launch the substeps: 
\itemize{
     \item peakpicking
     \item rT correction
     \item alignment
     \item fill missing peaks
}
Some parameters have a default parameter. This is the case for : 
\itemize{
     \item verboseColumns is TRUE cause we want all the infos 
         exported by XCMS
     \item binSize is 1 cause i find it sufficient & 
         1 is too high to perform a well alignment of the spectras
     \item sampleGroups is equal to the length of the provided files
         I prefer to apply some rules to filter later & that will allow
         the possibility to change this anytime instead of re-run 
         the entire process
     \item minFraction is the minimum integer possible for the same 
         reason than `sampleGroups`
     \item minSamples is at 1, same reason than `sampleGroups`
     \item maxFeatures is at 500 cause the workflow will probably 
         need to deal with some huge datasets
}
}
\examples{
\dontrun{
    mzxml_files <- c(
        system.file("testdata", "200204PLF_QC01_pos_filtered.mzML", 
             package = "workflow.lipido"), 
        system.file("testdata", "200204PLF_QC02_pos_filtered.mzML", 
             package = "workflow.lipido"))
    cwt_params <- xcms::CentWaveParam(
        ppm = 30, 
        peakwidth = c(4, 39), 
        snthresh = 6.5, 
        prefilter = c(2, 815), 
        mzdiff = .041, 
        noise = 0, 
        firstBaselineCheck = FALSE)
    obw_params <- xcms::ObiwarpParam() 
    pd_params <- xcms::PeakDensityParam(
         bw = 5, 
         binSize = .010
    )
    ms_process(mzxml_files, cwt_params, obw_params, pd_params)
}
}
