% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiwarp.R
\name{obiwarp}
\alias{obiwarp}
\title{Alignment: Retention time correction methods.}
\usage{
obiwarp(z, cntr, cntr_pr, parms)
}
\arguments{
\item{z}{XCMSnExp object of the file to adjust time retentions}

\item{cntr}{XCMSnExp object of the file used as reference}

\item{cntr_pr}{a profile matrix of the file used as reference (rows 
representing scans, columns equally spaced m/z values).}

\item{parms}{ObiwarpParam object created by xcms}
}
\value{
vector with the rT adjusted of the file according the rT of the 
     file used as reference
}
\description{
perform retention time correction (alignment) between chromatograms of 
     different samples. 
the alignment is based on the complete mz-rt data. This method does not 
     require any identified peaks or defined features. 
This function comes from the XCMS package and only the function `.obiwarp` 
     was imported in order to use & export the parallelisation infos
}
\seealso{
xcms:::adjustRtime
}
\author{
Johannes Rainer
}
