% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_db_eic}
\alias{plot_db_eic}
\title{Plot DB EIC}
\usage{
plot_db_eic(db, row_id)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{row_id}{`numeric(1)` ROWID of the annotation line in the database}
}
\value{
`plotly`
}
\description{
Plot the EIC for all the sample for a basepeak mass recorded in the database
The line dashed correspond to the area not integrated & the line colored the
retention time range where integrated by XCMS.
It contains a special behavior when the mouse hover a trace : it will display
 all the hovertext of all traces in a unique textbox allowing the user to
  differentiate all the y coordinates of the traces in one shot
If available it will use the retention time corrected in the slot
`scantime_corrected` added by the function `obiwarp`
}
\examples{
\dontrun{
plot_db_eic(db, row_id = 1)
}
}
