% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_empty_ms}
\alias{plot_empty_ms}
\title{Construct empty MS}
\usage{
plot_empty_ms(title = "Mass Spectra", y_title = "Intensity")
}
\arguments{
\item{title}{`character(1)` title of the plot}

\item{y_title}{`character(1)` title of the Y axis}
}
\value{
`plotly` object
}
\description{
Construct an empty MS
}
