% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_empty_peak_spot}
\alias{plot_empty_peak_spot}
\title{Plot empty MS map}
\usage{
plot_empty_peak_spot(
  title = "MS map",
  xaxis_title = "Retention time",
  yaxis_title = "m/z"
)
}
\arguments{
\item{title}{`character(1)` title of the plot}

\item{xaxis_title}{`character(1)` title of the x axis}

\item{yaxis_title}{`character(1)` title of the y axis}
}
\value{
`plotly` object
}
\description{
Plot an empty MS map with custom axis
}
