% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_peak_spot}
\alias{plot_peak_spot}
\title{Plot MS map}
\usage{
plot_peak_spot(
  db,
  annotation_filter = "all",
  int_threshold = 0,
  type = "MS map"
)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{annotation_filter}{`character(1)` annotation to filter (can be
"no annotated", "annotated" or "all")}

\item{int_threshold}{`numeric(1)` intensity threshold}

\item{type}{`character(1)` "MS map" or "Kendrick plot"}
}
\value{
`plotly`
}
\description{
Plot a MS map or Kendrick plot
Each basepeak are plotted with their annotations
Hover a trace will show all popup for all basepeak with the same group ID (
same compound flagged by CAMERA)
Possibility to trace the "MS map" (m/z fct(rT)) or a Kendrick plot
Each trace will contain the EIC ID in their `customdata` slot
}
\examples{
\dontrun{
plot_peak_spot(db)
}
}
