% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_raw_tic}
\alias{plot_raw_tic}
\title{Plot TICs}
\usage{
plot_raw_tic(raw_files, msaccess, polarity = "positive")
}
\arguments{
\item{raw_files}{`character vector` filepaths to the raw files}

\item{msaccess}{`character(1)` filepath to the msaccess executable file}

\item{polarity}{`character(1)` should be "positive" or "negative" only}
}
\value{
`plotly`
}
\description{
Plot TICs from raw files (not converted). It use in background the msaccess
executable from ProteoWizard
}
\examples{
\dontrun{
plot_raw_tic(
   c("SPRING/inst/testdata/220221CCM_global_POS_01_ssleu_filtered.mzML",
     "SPRING/inst/testdata/220221CCM_global_POS_01_ssleu_filtered.mzML"),
    "pwiz/msaccess.exe",
    "positive"
)
}
}
