% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthData.R
\name{qstepcdf}
\alias{qstepcdf}
\title{Inverse function of empirical cumulative distribution function (ecdf)}
\usage{
qstepcdf(p, empf, interval, tol = 0.001, maxiter = 100)
}
\arguments{
\item{p}{probability (between 0 and 1)}

\item{empf}{empirical cdf or any quantile function.}

\item{interval}{find a solution only within this interval. a vector containing two end points of the interval.}

\item{tol}{the desired accuracy (convergence tolerance).}

\item{maxiter}{the maximum number of iterations for \code{uniroot.all}.}
}
\value{
A numeric scalar: the count value corresponding to probability \code{p}
  under the empirical CDF.
}
\description{
Inverse function of empirical cumulative distribution function (ecdf)
}
\examples{
### This is an internal function used in synthData_from_ecdf.
}
