% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{reintegrate_ann}
\alias{reintegrate_ann}
\title{Reintegrate compound}
\usage{
reintegrate_ann(db, cpd_name, rtmin, rtmax)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{cpd_name}{`character(1)` compound name}

\item{rtmin}{`numeric(1)` rT born min}

\item{rtmax}{`numeric(1)` rT born max}
}
\value{
`DataFrame list` with items :
\itemize{
    \item ann `DataFrame` each line correspond to a compound found
    with the columns:
    \itemize{
        \item pcgroup_id `integer` group ID
        \item basepeak_group_id `integer` EIC ID
        \item class `character` cpd class
        \item name `character` name
        \item formula `character` chemical formula
        \item adduct `character` adduct form
        \item ion_formula `character` ion chemical formula
        \item rtdiff `numeric` retention time difference between the measured
         & the expected
        \item rt `numeric` retention time measured meanned accross the
        samples
        \item rtmin `numeric` born min of retention time measured accross the
        samples
        \item rtmax `numeric` born max of the retention time measured accross
         the samples
        \item nsamples `integer` number of samples where the compound was
        found
        \item best_score `numeric` best isotopic score seen
        \item best_deviation_mz `numeric` best m/z deviation seen
        \item best_npeak `integer` best number of isotopologues found
        \item ... `integer` a column for each sample which contain the
        spectra ID
    }
    \item spectras `DataFrame`, each line correspond to a peak annotated with
    its corresponding theoretical peak or the theoretical peak missed,
    with the columns :
        \itemize{
        \item spectra_id `integer` spectra ID
        \item feature_id `integer` feature ID
        \item mz `numeric` m/z
        \item int `numeric` area integrated
        \item abd `numeric` relative abundance
        \item ion_id_theo `integer` ignore
        \item mz_theo `numeric` theoretical m/z
        \item abd_theo `numeric` theoretical relative abundance
        \item iso_theo `character` theoretical isotopologue annotation
    }
    \item spectra_infos `DataFrame`, each line correspond to a spectra
    annotated, with the columns :
    \itemize{
        \item spectra_id `integer` spectra ID
        \item score `numeric` isotopic score observed
        \item deviation_mz `numeric` m/z deviation observed
        \item npeak `integer` number of isotopologue annotated
        \item basepeak_mz `numeric` m/z of the basepeak annotated
        \item basepeak_int `numeric` area of the basepeak annotated
        \item sum_int `numeric` cumulative sum off all the area of the
        isotopologues annotated
        \item rt `numeric` retention time
    }
    \item peakgroups `DataFrame`, each line correspond to a group annotated
    by XCMS and CAMERA
    \itemize{
        \item group_id `integer` group ID
        \item pcgroup_id `integer` pcgroup ID
        \item adduct `character` adduct annotation by CAMERA (NULL if absent)
        \item cluster_id `integer` cluster ID
        \item iso `character` could be "M" or "M+*"
        \item mzmed `float` m/z median computed by XCMS
        \item mzmin `float` m/z median born min (not the m/z born min !!)
        \item mzmax `float` m/z median born max (not the m/z born max !!)
        \item rtmed `float` rT median computed by XCMS
        \item rtmin `float` rT median born min (not the rT born min !!)
        \item rtmax `float` rT median born max (not the rT born max !!)
        \item npeaks `integer` number of peaks grouped accross samples
        \item ... `integer` a column for each sample which contain the
        feature ID (row ID from the peaktable)
    }
    \item peaks `DataFrame`, peaktable from XCMS
    \itemize{
        \item mz `float` m/z
        \item mzmin `float` m/z born min
        \item mzmax `float` m/z born max
        \item rt `float` rT
        \item rtmin `float` rT born min
        \item rtmax `float` rT born max
        \item into `float` area of the peak
        \item intb `float` area of the peak above baseline
        \item maxo `float` maximum intensity
        \item sn `float` signal/noise
        \item egauss `float` ignore
        \item mu `float` ignore
        \item sigma `float` ignore
        \item h `float` ignore
        \item f `integer` ID of the ROI
        \item dppm `float` ppm deviation
        \item scale `integer` width of the wave used for the peak detection
        \item scpos `integer` scan ID
        \item scmin `integer` scan ID born min of the wave detection
        \item scmax `integer` scan ID born max of the wave detection
        \item lmin `integer` scan ID after extension of the scmin
        \item lmax `integer` scan ID after extension of the scmax
    }
}
}
\description{
Reintegrate compound by giving a new rT range
It will reintegrate all isotopologues for all files for all ions for this
 compound. It will recreate spectra, peakgroups & all new entries.
}
