% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{resolve_conflicts}
\alias{resolve_conflicts}
\title{Resolve annotations conflicts}
\usage{
resolve_conflicts(ann)
}
\arguments{
\item{ann}{the annotation dataframe obtained 
from the function `annotate_peaklists`}
}
\value{
the annotation dataframe with all annotations conflicts solved
}
\description{
Resolve annotations conflicts which corresponds to multiple annotations 
     for the same group of peaks
it will priorize those which were found in the maximum of sample 
     then the annotations with the maximum of different adducts
     then those with the minimum of retention time compared to theoretical
     finally by the isotopic score
}
