% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run Shiny app}
\usage{
run_shiny_app(pwiz_dir = "pwiz", log_file_path = NULL)
}
\arguments{
\item{log_file_path}{`character(1)` path to the file were to write the stderr
& stdout}

\item{pwiz}{`character(1)` path to the pwiz directory}
}
\description{
Run a shiny application in the default web browser
}
\examples{
\dontrun{
     run_shiny_app()
}
}
