% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{split_ms_file}
\alias{split_ms_file}
\title{Split MS file}
\usage{
split_ms_file(ms_file, exp_polarity, overwrite = TRUE)
}
\arguments{
\item{ms_file}{`OnDiskMSnExp` MSnbase object}

\item{exp_polarity}{`character(1)` expected polarity ("positive" or
"negative")}

\item{overwrite}{`logical(1)` should overwrite the file if needed ?}
}
\value{
`character(1)` filepath
}
\description{
Split MS file to keep only the scans in the desired polarity & rewrite it
in a new file
}
