% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{spring.generator}
\alias{spring.generator}
\title{Generate multivariate data}
\usage{
spring.generator(model, design)
}
\description{
Generate multivariate data
}
\examples{
\dontrun{##  Basic simulation settings with no particular structure
k <- 10
p <- 20
q <- 3

## direct coefficients
Omega.xy <- Matrix(0,p,q)
Omega.xy[sample(1:(p*q),k)] <- sample(c(-1,1),k,rep=T)

## covariance of the noise (residual)
cor <- 0.5
R <- toeplitz(cor^(1:q-1))

## turn this into a "spring" model
true.model <- new("model",
                  coef.direct = Omega.xy,
                  coef.regres = Matrix(- Omega.xy \%*\% R),
                  covariance  = Matrix(R),
                  intercept   = Matrix(rep(0,q),q,1))

## design matrix
nb.samples <- 100
X <- matrix(rnorm(nb.samples * p), nb.samples, p)

## draw some data with a given design matrix
dataset <- spring.generator(true.model, X)
}

}
