% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SynthData}
\alias{SynthData}
\alias{SynthData2}
\title{Synthetic count data}
\format{
\code{SynthData} is an object of class \code{matrix} with 500 rows and 200 columns.

An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 100 columns.
}
\source{
Yoon, Gaynanova and Müller (2019) Microbial Networks in SPRING - Semi-parametric Rank-Based Correlation and Partial Correlation Estimation for Quantitative Microbiome Data. \emph{Frontiers in Genetics.} 10:516. \url{doi:10.3389/fgene.2019.00516}
}
\usage{
data(SynthData)

data(SynthData2)
}
\description{
SynthData and SynthData2 were generated using empirical cdf of American Gut Project Data.
SynthData has scale_free-type-graph structure of size 500 rows and 200 columns, and SynthData2 has cluster-type-graph structure of size 1000 rows and 100 columns.
}
\keyword{datasets}
