% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SynthData2}
\alias{SynthData2}
\title{This is a Synthetic count data to be included in this package. Generated using empirical cdf of American Gut Project Data with cluster graph type.}
\format{A matrix of dimension with 1000 rows and 100 columns.}
\usage{
SynthData2
}
\description{
This is a Synthetic count data to be included in this package. Generated using empirical cdf of American Gut Project Data with cluster graph type.
}
\keyword{datasets}
