% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{annotate_peaklists}
\alias{annotate_peaklists}
\title{Annotate peaklists}
\usage{
annotate_peaklists(ms_files, ann_params, show_pb, sigma = 6, perfwhm = 0.6)
}
\arguments{
\item{ms_files}{XCMSnExp}

\item{ann_params}{AnnotationParameter object}

\item{show_pb}{boolean print progression bar}

\item{sigma}{ignore}

\item{perfwhm}{ignore}
}
\value{
a XCMSnExp
}
\description{
Annotate peaklists from an XCMSnExp with grouping information
}
