% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_eic}
\alias{db_get_eic}
\title{Get EIC from data}
\usage{
db_get_eic(db, eic_id)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{eic_id}{`numeric(1)` eic ID in database, it corresponds to the row ID
for the annotation table in the database}
}
\value{
`DataFrame` with a variable number of columns
 (one sample = one column):
}
\description{
Get all the EIC for all the files in the database for ONE basepeak mass
}
