% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_record_eics}
\alias{db_record_eics}
\title{Record EICs}
\usage{
db_record_eics(db, pb_fct = NULL)
}
\arguments{
\item{db}{`SQLiteConnection`}
}
\description{
Record EICs for each basepeak for each group ID for each sample
It will be faster than reloading each raw file from database & retrace the
corresponding EIC
}
