% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hugeKmb}
\alias{hugeKmb}
\title{Internal wrapper function to implement rank-based correlation to huge.mb function in "huge" package.}
\usage{
hugeKmb(data, lambda, type = "trunc", sym = "or", verbose = TRUE)
}
\arguments{
\item{data}{n by p matrix data. usually through pulsar, data will receive subsamples.}

\item{lambda}{a vector of lambda values}

\item{type}{a type of variables. "trunc" is default.}

\item{sym}{"or" is the symmetrizing rule of the output graphs. If sym = "and", the edge between node i and node j is selected ONLY when both node i and node j are selected as neighbors for each other. If sym = "or", the edge is selected when either node i or node j is selected as the neighbor for each other. The default value is "or". (refer to huge manual)}

\item{verbose}{default is TRUE.}
}
\value{
\code{hugeKmb} returns a data.frame containing
\itemize{
     \item{beta: }
     \item{path: }{a list of}
     \item{df: }
}
}
\description{
Internal wrapper function to implement rank-based correlation to huge.mb function in "huge" package.
}
