% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mclr}
\alias{mclr}
\title{Modified central log ratio (mclr) transformation}
\usage{
mclr(dat, base = exp(1), tol = 1e-16, eps = NULL, atleast = 1)
}
\arguments{
\item{dat}{raw count data or compositional data (n by p) does not matter.}

\item{base}{exp(1) for natural log}

\item{tol}{tolerance for checking zeros}

\item{eps}{epsilon in eq (2) of the paper "Yoon, Gaynanova, M\"{u}ller (2019), Frontiers in Genetics". positive shifts to all non-zero compositions. Refer to the paper for more details. eps = absolute value of minimum of log ratio counts plus c.}

\item{atleast}{default value is 1. Constant c which ensures all nonzero values to be strictly positive. default is 1.}
}
\value{
\code{mclr} returns a data matrix of the same dimension with input data matrix.
}
\description{
Modified central log ratio (mclr) transformation
}
\examples{
data(QMP)
RMP <- QMP/rowSums(QMP)
mclr_RMP <- mclr(RMP)

}
