% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ms_process}
\alias{ms_process}
\title{Execute The step "Process"}
\usage{
ms_process(
  mzxml_files,
  cwt_params,
  obw_params,
  cores = parallel::detectCores(),
  show_pb = TRUE
)
}
\arguments{
\item{mzxml_files}{vector of filepaths to the files}

\item{cwt_params}{CentWaveParam object created by xcms, 
if not given it will launch the parameters by default}

\item{obw_params}{ObiwarpParam object created by xcms, 
if not given it will launch the parameters by default}

\item{cores}{numeric number of cores to use for the peakpicking 
(one core = one file peakpicked)}

\item{show_pb}{boolean print a progress bar or not}
}
\value{
XCMSnExp object
}
\description{
Execute the step "Process" of the workflow.
it launch the substeps: 
\itemize{
     \item peakpicking
     \item rT correction
     \item alignment
     \item fill missing peaks
}
}
\examples{
\dontrun{
    mzxml_files <- c(
        system.file("testdata", "200204PLF_QC01_pos_filtered.mzML", 
             package = "workflow.lipido"), 
        system.file("testdata", "200204PLF_QC02_pos_filtered.mzML", 
             package = "workflow.lipido"))
    cwt_params <- xcms::CentWaveParam(
        ppm = 30, 
        peakwidth = c(4, 39), 
        snthresh = 6.5, 
        prefilter = c(2, 815), 
        mzCenterFun = "wMean", 
        integrate = 1, 
        mzdiff = .041, 
        fitgauss = FALSE, 
        noise = 0, 
        verboseColumns = TRUE, 
        firstBaselineCheck = FALSE)
    obw_params <- xcms::ObiwarpParam(
        binSize = .1,
        centerSample = integer(),
        response = 1L,
        distFun = "cor_opt",
        gapInit = .3,
        gapExtend = 2.4,
        factorDiag = 2,
        factorGap = 1,
        localAlignment = FALSE,
        initPenalty = 0
    ) 
    ms_process(mzxml_files, cwt_params)
}
}
