% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiwarp.R
\name{obiwarp}
\alias{obiwarp}
\title{Alignment: Retention time correction methods.}
\usage{
obiwarp(ms_files, obw_params, operator, show_pb)
}
\arguments{
\item{ms_files}{list of XCMSnExp}

\item{obw_params}{ObiwarpParam object created by xcms}

\item{operator}{function to use for parallelization (`\%dopar\%`) 
or not (`\%do\%`)}

\item{show_pb}{boolean indicate if we print a progress bar or not}
}
\value{
XCMSnExp which regroup all the files in one object 
     with the rT corrected
}
\description{
perform retention time correction (alignment) between chromatograms of 
     different samples. 
the alignment is based on the complete mz-rt data. This method does not 
     require any identified peaks or defined features. 
This function comes from the XCMS package and only the function `.obiwarp` 
     was imported in order to use & export the parallelisation infos
It force to use the sample with the most peak as the center sample
}
\seealso{
xcms:::adjustRtime
}
