% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion-class.R
\docType{methods}
\name{plot,criterion-method}
\alias{plot,criterion-method}
\alias{plot.criterion}
\title{Plot method for object criterion}
\usage{
\\S4method{plot}{criterion}(x, y, xvar="lambda1",
log.scale=ifelse(xvar=="lambda1",TRUE,FALSE), plot=TRUE, ...)
}
\arguments{
\item{x}{an object with class criterion}

\item{y}{used for S4 compatibility.}

\item{xvar}{a character among \code{"lambda1"}, \code{"fraction"}
or \code{"df"} for the x-axis (only \code{lambda1} is available
when the criterion object come from a call to \code{cv.spring}).}

\item{log.scale}{a boolean to put the x-axis on a log-scale.}

\item{plot}{is \code{TRUE}, a plot is produce. Otherwise a
\code{ggplot} object is sent back.}
}
\value{
a invisible ggplot object.
}
\description{
A plotting method for the class \code{criterion}.
}
\seealso{
\code{\linkS4class{criterion}}.
}
