% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QMP}
\alias{QMP}
\title{Quantitative Microbiome Project data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 106 rows and 91 columns.
}
\source{
Yoon, Gaynanova and Müller (2019) Microbial Networks in SPRING - Semi-parametric Rank-Based Correlation and Partial Correlation Estimation for Quantitative Microbiome Data. \emph{Frontiers in Genetics}. 10:516. \url{doi:10.3389/fgene.2019.00516}

Vanderputte et al. (2017) Quantitative microbiome profiling links gut community variation to microbial load. \emph{Nature}. 551: 507-511. \url{doi:10.1038/nature24460}
}
\usage{
QMP
}
\description{
The data containing quantitative microbiome count data of dimenstion 106 samples/subjects (in rows) and 91 OTUs (in columns). The raw dataset is pruned the taxa present less than 30\% of samples and final dataset contains only healthy subjects from two cohorts: Study cohort and Disease cohort.
}
\keyword{datasets}
