% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spring-package.R
\docType{package}
\name{spring-package}
\alias{spring-package}
\title{SPRING}
\description{
This package is designed to fit.
}
\references{
Julien Chiquet, Tristan Mary-Huard, Stephane
Robin. Structured regularization for conditional Gaussian
graphical models. Statistics and Computing, 4 mai 2016.
doi:10.1007/s11222-016-9654-1.
}
\author{
Julien Chiquet \email{julien.chiquet@genopole.cnrs.com}
}
