\docType{package}
\name{spring-package}
\alias{spring-package}
\title{SPRING}
\description{
This package is designed to fit.
}
\section{Features}{

}

\section{Algorithm}{

}

\section{Technical remarks}{

}
\author{
Julien Chiquet \email{julien.chiquet@genopole.cnrs.com}
}
\references{
Julien Chiquet, Tristan Mary-Huard, Stephane Robin.
Structured Regularization for conditional Gaussian
Graphical models. arXiv preprint, \url{}
}

