% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRING.R
\name{SPRING}
\alias{SPRING}
\title{SPRING main function
Semi-Parametric Rank-based approach for INference in Graphical model. SPRING follows the neighborhood selection methodology outlined in "mb" Meinshausen and Buhlmann and}
\usage{
SPRING(data, quantitative = FALSE, method = "mb",
  lambda.min.ratio = 0.01, nlambda = 50, lambdaseq = NULL,
  seed = 10010, ncores = 2, thresh = 0.1, subsample.ratio = 0.8,
  rep.num = 50)
}
\arguments{
\item{data}{n by p matrix of microbiome data. Each row represents each subject/sample and each column represents each OTU (operational taxonomic unit).}

\item{quantitative}{default is FALSE, which means input "data" is compositional data, which will be normalized using mclr transformation within a function. If TRUE, it means "quantitative" counts are input and no normalization will be applied.}

\item{method}{graph estimation methods. Currently, only "mb" method is available.}

\item{lambda.min.ratio}{default is 0.01}

\item{nlambda}{default is 50.}

\item{lambdaseq}{a sequence of decreasing positive numbers to control the regularization. Users can specify a sequence to override the default sequence.}

\item{seed}{the seed for subsampling}

\item{ncores}{number of cores to use subsampling}

\item{thresh}{threshold for StARS selection criterion. 0.1 is recommended.}

\item{subsample.ratio}{0.8 is default. The recommended values are 10*sqrt(n)/n for n > 144 or 0.8 otherwise.}

\item{rep.num}{the repetition number of subsampling}
}
\value{
\code{SPRING} returns a data.frame containing
}
\description{
SPRING main function
Semi-Parametric Rank-based approach for INference in Graphical model. SPRING follows the neighborhood selection methodology outlined in "mb" Meinshausen and Buhlmann and
}
\examples{
library(SpiecEasi)
data(amgut1.filt)
fit <- SPRING(data = amgut1.filt[, 1:10], quantitative = FALSE, nlambda = 20, rep.num = 10)
}
