% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRING.R
\name{SPRING}
\alias{SPRING}
\title{Semi-Parametric Rank-based approach for INference in Graphical model (SPRING)}
\usage{
SPRING(
  data,
  quantitative = FALSE,
  method = c("mb"),
  lambda.min.ratio = 0.01,
  nlambda = 20,
  lambdaseq = exp(seq(log(0.6), log(0.6 * lambda.min.ratio), length.out = nlambda)),
  seed = 10010,
  ncores = 1,
  thresh = 0.1,
  subsample.ratio = 0.8,
  rep.num = 20,
  Rtol = 1e-06,
  verbose = TRUE,
  Rmethod = "approx",
  use.nearPD = TRUE,
  nu = 0.001,
  ratio = 0.9
)
}
\arguments{
\item{data}{n by p matrix of microbiome count data, either quantitative or compositional counts. Each row represents each subject/sample and each column represents each OTU (operational taxonomic unit).}

\item{quantitative}{default is FALSE, which means input "data" is compositional data, which will be normalized using mclr transformation within a function. If TRUE, it means "quantitative" counts are input and no normalization will be applied.}

\item{method}{graph estimation method. Currently only \code{"mb"} (Meinshausen-Bühlmann neighborhood selection) is available.}

\item{lambda.min.ratio}{ratio of the smallest to largest value in the lambda sequence. Default is 0.01.}

\item{nlambda}{number of lambda values in the regularization sequence. Default is 20.}

\item{lambdaseq}{a sequence of decreasing positive numbers to control the regularization. The default sequence has 20 values generated to be equally spaced on a logarithmic scale from 0.6 to 0.006. Users can specify a sequence to override the default sequence. If user specify as "data-specific", then the lambda sequence will be generated using estimated rank-based correlation matrix from data.}

\item{seed}{the seed for subsampling.}

\item{ncores}{number of cores to use for subsampling. The default is 1.}

\item{thresh}{threshold for StARS selection criterion. 0.1 is recommended (default). The smaller threshold returns sparser graph.}

\item{subsample.ratio}{0.8 is default. The recommended values are 10*sqrt(n)/n for n > 144 or 0.8 otherwise.}

\item{rep.num}{the repetition number of subsampling for StARS edge stability selection. The default value is 20.}

\item{Rtol}{Desired accuracy when calculating the solution of bridge function in latentcor. Default is 1e-6.}

\item{verbose}{If \code{verbose = FALSE}, tracing information printing for HUGE (High-dimensional Undirected Graph Estimation) is disabled. The default value is TRUE.}

\item{Rmethod}{The calculation method of latent correlation. Either \code{"approx"} or \code{"original"}. If \code{Rmethod = "approx"}, multilinear approximation method is used, which is much faster than the original method. If \code{Rmethod = "original"}, optimization of the bridge inverse function is used. The default is \code{"approx"}.}

\item{use.nearPD}{Logical indicator. \code{use.nearPD = TRUE} gets nearest positive definite matrix for the estimated latent correlation matrix with shrinkage adjustment by \code{nu}. Output \code{R} is the same as \code{Rpointwise} if \code{use.nearPD = FALSE}. Default value is \code{TRUE}.}

\item{nu}{Shrinkage parameter for the correlation matrix, must be between 0 and 1. Guarantees that the minimal eigenvalue of the returned correlation matrix is greater or equal to \code{nu}. The default (recommended) value is 0.001.}

\item{ratio}{When \code{Rmethod = "approx"}, specifies the boundary value for multilinear interpolation, must be between 0 and 1. The default (recommended) value is 0.9. Ignored when \code{Rmethod = "original"}.}
}
\value{
\code{SPRING} returns a list containing
\itemize{
      \item{output: }{Output results of \code{pulsar::pulsar} based on StARS criterion. It contains:}
              \itemize{
                       \item{merge: } a list of length \code{nlambda} and each element of list contains a matrix of edge selection probability. Each lambda value, this edge selection probability is calculated across \code{rep.num}.
                       \item{summary: } the summary statistic over \code{rep.num} graphs at each value of lambda
                       \item{opt.index: } index (along the path) of optimal lambda selected by the criterion at the desired threshold. Will return \eqn{0} if no optimum is found or \code{NULL} if selection for the criterion is not implemented.
                       \item{criterion: } we use StARS for our stability criterion.
              }
      \item{fit: }{Output results of \code{pulsar::refit} function. It contains:}
      \itemize{
              \item{est: } a list containing
              \itemize{
                       \item{beta: } Estimates of beta coefficient matrices (of size p by p) by "mb" method on the whole data at each of whole lambda sequence value.
                       \item{path: } Estimates of precision matrix (of size p by p) on the whole data at each of whole lambda sequence value.
              }
              \item{refit: } final estimates of precision matrix (of size p by p).
              }
      \item{lambdaseq: }{lambda sequence used in the analysis}
}
}
\description{
SPRING follows the neighborhood selection methodology outlined in "mb" method (Meinshausen and Buhlmann (2006)).
}
\examples{
rm(list = ls())
library(SPRING)

# Load the synthetic count data
data("QMP") # n = 1000 and p = 100 synthetic dataset

# SPRING on Synthetic Data, when assuming the data as quantitative counts.
# The same setting used in Yoon et al. (2019) Frontiers in Genetics.
\dontrun{
# This takes around 23 minutes.
fit.spring <- SPRING(QMP, quantitative = TRUE, lambdaseq = "data-specific",
                     nlambda = 50, seed = 10010, ncores = 2, rep.num = 50)
}

# SPRING on Compositional data. Row sums are scaled to 1. Then, mclr-transformation will be applied.
\dontrun{
compoData <- QMP/rowSums(QMP)
fit.spring <- SPRING(compoData, quantitative = FALSE, lambdaseq = "data-specific",
                     nlambda = 10, rep.num = 10)
}
}
\references{
Meinshausen N. and Buhlmann P. (2006) \href{https://projecteuclid.org/download/pdfview_1/euclid.aos/1152540754}{"High-dimensional graphs and variable selection with the lasso"}, \emph{The Annals of Statistics}, Vol 34, No. 3, 1436 - 1462.

Yoon G., Gaynanova I. and Müller C. (2019) \href{https://www.frontiersin.org/articles/10.3389/fgene.2019.00516/full}{"Microbial Networks in SPRING - Semi-parametric Rank-Based Correlation and Partial Correlation Estimation for Quantitative Microbiome Data"}, \emph{Frontiers in Genetics}, 10:516.
}
