% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SynthData}
\alias{SynthData}
\title{This is a Synthetic count data to be included in this package. Generated using empirical cdf of American Gut Project Data with scale_free graph type.}
\format{A matrix of dimension with 500 rows and 200 columns.}
\usage{
SynthData
}
\description{
This is a Synthetic count data to be included in this package. Generated using empirical cdf of American Gut Project Data with scale_free graph type.
}
\keyword{datasets}
